requirements.txt
streamlit
opencv-python
mediapipe
tensorflow
numpy
scikit-learn
speech-recognition
pyttsx3
googletrans==4.0.0rc1
pyaudio
